Move increment_field to models feature
